progress on refactoring models
