+++ .gitignore
@@ -1,3 +1,4 @@
+.idea/
 *.png
 *.pkl
 *.gif
